+++ src/main/kotlin/pt/isel/ls/sessions/repository/data/game/GameMemoryDB.kt
@@ -6,14 +6,14 @@ import java.util.concurrent.atomic.AtomicInteger
 
 class GameMemoryDB : GameDB {
 
-    private val gameMap = ConcurrentHashMap<Int, Game>()
+    private val gameMap = ConcurrentHashMap<UInt, Game>()
     private var nextGameId = AtomicInteger(1)
 
-    override fun createGame(name: String, developer: String, genres: List<Genres>): Int? =
+    override fun createGame(name: String, developer: String, genres: List<Genres>): UInt? =
         if (gameMap.any { it.value.name == name })
             null
         else {
-            val gid = nextGameId.getAndIncrement()
+            val gid = nextGameId.getAndIncrement().toUInt()
             gameMap[gid] = Game(gid, name, developer, genres)
             gid
         }
@@ -23,7 +23,7 @@ class GameMemoryDB : GameDB {
     }.filter { game -> game.developer == developer }.toList()
 
 
-    override fun getGameById(gid: Int): Game? = gameMap[gid]
+    override fun getGameById(gid: UInt): Game? = gameMap[gid]
 
     override fun reset(): Unit = run {
         gameMap.clear()
